fix: adding lombok support + update java version

+++ src/main/java/com/renan/booksalesonline/adapters/repositories/mappers/LanguageEntityMapper.java
@@ -32,6 +32,6 @@ public class LanguageEntityMapper {
         if (language == null)
             return null;
 
-        return new LanguageEntity(language.getId(), language.getName());
+        return new LanguageEntity((int) language.getId(), language.getName());
     }
 }
